+++ src/main/java/io/iworkflow/core/exceptions/LongPollTimeoutException.java
@@ -1,5 +1,7 @@
 package io.iworkflow.core.exceptions;
 
+import io.iworkflow.core.ClientSideException;
+
 public class LongPollTimeoutException extends ClientSideException {
     public LongPollTimeoutException(
             final ClientSideException exception) {
