+++ src/main/java/io/iworkflow/core/exceptions/WorkflowAlreadyStartedException.java
@@ -1,5 +1,7 @@
 package io.iworkflow.core.exceptions;
 
+import io.iworkflow.core.ClientSideException;
+
 public class WorkflowAlreadyStartedException extends ClientSideException {
     public WorkflowAlreadyStartedException(
             final ClientSideException exception) {
