+++ src/main/java/io/iworkflow/core/exceptions/WorkflowNotExistsOrOpenException.java
@@ -1,5 +1,7 @@
 package io.iworkflow.core.exceptions;
 
+import io.iworkflow.core.ClientSideException;
+
 public class WorkflowNotExistsOrOpenException extends ClientSideException {
     public WorkflowNotExistsOrOpenException(
             final ClientSideException exception) {
